requirements.txt
tensorflow==2.11.0
keras==2.11.0
pandas==1.3.5
beautifulsoup4==4.12.2
google-auth==2.9.0
pandas-gbq==0.17.6